Cast UIControlState values to unsigned long and use the unsigned long string format specifier (%lu) when writing them to a string. NSUInteger is defined as unsigned int on 32-bit and unsigned long on 64-bit. This fix allows the code to build on 32- or 64-bit architectures without warnings.

+++ FTWButton/FTWButton.m
@@ -457,25 +457,25 @@
 
 - (void) setValue:(id)value inDictionary:(NSMutableDictionary*)dictionary forControlState:(UIControlState)controlState {
 	if (value) {
-		[dictionary setValue:value forKey:[NSString stringWithFormat:@"%d",controlState]];
+		[dictionary setValue:value forKey:[NSString stringWithFormat:@"%lu", (unsigned long)controlState]];
 	} else {
-		[dictionary removeObjectForKey:[NSString stringWithFormat:@"%d",controlState]];
+		[dictionary removeObjectForKey:[NSString stringWithFormat:@"%lu", (unsigned long)controlState]];
 	}
 	[self configureViewForControlState:[self currentControlState]];
 }
 
 - (id) getValueFromDictionary:(NSMutableDictionary*)dictionary forControlState:(UIControlState)controlState {
-	if ([dictionary valueForKey:[NSString stringWithFormat:@"%d",controlState]]) {
-		return [dictionary valueForKey:[NSString stringWithFormat:@"%d",controlState]];
+	if ([dictionary valueForKey:[NSString stringWithFormat:@"%lu", (unsigned long)controlState]]) {
+		return [dictionary valueForKey:[NSString stringWithFormat:@"%lu", (unsigned long)controlState]];
 	}
 	
 	
-	if ((controlState & UIControlStateSelected) && [dictionary valueForKey:[NSString stringWithFormat:@"%d",UIControlStateSelected]]) {
-		return [dictionary valueForKey:[NSString stringWithFormat:@"%d",UIControlStateSelected]];
-	} else if ((controlState & UIControlStateHighlighted) && [dictionary valueForKey:[NSString stringWithFormat:@"%d", UIControlStateHighlighted]]) {
-		return [dictionary valueForKey:[NSString stringWithFormat:@"%d", UIControlStateHighlighted]];
+	if ((controlState & UIControlStateSelected) && [dictionary valueForKey:[NSString stringWithFormat:@"%lu", (unsigned long)UIControlStateSelected]]) {
+		return [dictionary valueForKey:[NSString stringWithFormat:@"%lu", (unsigned long)UIControlStateSelected]];
+	} else if ((controlState & UIControlStateHighlighted) && [dictionary valueForKey:[NSString stringWithFormat:@"%lu", (unsigned long)UIControlStateHighlighted]]) {
+		return [dictionary valueForKey:[NSString stringWithFormat:@"%lu", (unsigned long)UIControlStateHighlighted]];
 	} else {
-		return [dictionary valueForKey:[NSString stringWithFormat:@"%d",UIControlStateNormal]];
+		return [dictionary valueForKey:[NSString stringWithFormat:@"%lu", (unsigned long)UIControlStateNormal]];
 	}
 }
 
@@ -491,7 +491,7 @@
 	if (controlState & UIControlStateSelected) {
 		selectedLabel.textColor = newTextColor;
 	}
-	[textColors setValue:newTextColor forKey:[NSString stringWithFormat:@"%d",controlState]];
+	[textColors setValue:newTextColor forKey:[NSString stringWithFormat:@"%lu", (unsigned long)controlState]];
 	[self setValue:newTextColor inDictionary:textColors forControlState:controlState];
 }
 
@@ -634,7 +634,7 @@
 }
 
 - (void) setInnerShadowRadius:(CGFloat)shadowRadius forControlState:(UIControlState)controlState {
-	[innerShadowRadii setValue:[NSNumber numberWithFloat:shadowRadius] forKey:[NSString stringWithFormat:@"%d",controlState]];
+	[innerShadowRadii setValue:[NSNumber numberWithFloat:shadowRadius] forKey:[NSString stringWithFormat:@"%lu", (unsigned long)controlState]];
 	[self setValue:[NSNumber numberWithFloat:shadowRadius] inDictionary:innerShadowRadii forControlState:controlState];
 }
